requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
streamlit>=1.22.0
pandas>=1.5.3
polars>=0.20.0
pyarrow>=14.0.1
python-dotenv>=1.0.0
requests>=2.28.2
psycopg2-binary>=2.9.5
PyGithub>=1.58.1
python-dateutil>=2.8.2
argparse>=1.4.0
pandasai>=2.4.2
matplotlib>=3.7.1
Pillow>=10.0.0
PyPDF2>=3.0.0
tabula-py>=2.9.0
langchain>=0.1.0
langchain-community>=0.0.15
pgvector>=0.2.4
dropbox>=11.36.0
sentence-transformers>=2.2.2
huggingface-hub>=0.20.1
langchain-postgres>=0.0.2
openrouter>=0.8.0
scikit-learn>=1.3.0
# New dependencies for enhanced RAG
nltk>=3.8.1
pytesseract>=0.3.10
pdf2image>=1.16.3
psutil>=5.9.5
tiktoken>=0.5.1
python-Levenshtein>=0.21.1 